requirements.txt
altair==5.4.1
anyio==4.5.2
asgiref==3.10.0
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.3.3
cycler==0.12.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
exceptiongroup==1.3.0
fonttools==4.60.1
gitdb==4.0.12
GitPython==3.1.45
gunicorn==23.0.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
narwhals==1.42.1
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==24.2
pandas==2.3.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
protobuf==5.29.5
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.24.11
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.35.1
requests==2.32.4
rich==13.9.4
rpds-py==0.20.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.44.0
streamlit==1.40.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
watchdog==4.0.2
wheel==0.45.1
zipp==3.20.2
